Remove `namePrefix` from kustomize overlays
The `namePrefix` looked nice and did not cause extra
difficulty in the past so I used it. It's now causing
a little difficulty and it's a better trade off to just
drop it.

The new `OpenStackDataPlaneDeployment` spec references
`nodeSets` by metadata name and install_yamls produces
them with a matching name so that the Deployment and
NodeSet objects work together by default. However, the
`namePrefix` was changing the name in the metadata but
not the spec. Rather than make the kustomize overlays
more complciated to update the spec accordingly it's
easier to just drop the `namePrefix`.

I.e. the following was getting produced which won't
work since it references the noded set by the wrong
name: `openstack-edpm != standard-openstack-edpm`.

```
apiVersion: dataplane.openstack.org/v1beta1
kind: OpenStackDataPlaneDeployment
metadata:
  name: standard-openstack-edpm
  namespace: openstack
spec:
  nodeSets:
  - openstack-edpm
---
apiVersion: dataplane.openstack.org/v1beta1
kind: OpenStackDataPlaneNodeSet
metadata:
  name: standard-openstack-edpm
  namespace: openstack
```