Enable Glance image conversion
Update docs to include creating a PVC for in the glance pod
for staging the conversion and document how to confirm the
running Glance configuration.

Update kustomize overlay to use the PVC and pass parameters
to customServiceConfig so that glance-director and web-download
are working input methods. Also, enable the image_conversion
plugin and set the output format to raw.

Update test script have an option (IMPORT_RAW) to import an
already raw image. Otherwise, if ceph is enabled it will import
it using glance-direct as a qcow2 and we can then expect the
image conversion plugin to convert it to raw.